Move most of codegen into basm-std
