+++ basm-std/src/platform/mod.rs
@@ -1,3 +1,5 @@
+#[cfg(all(not(test), feature = "codegen"))]
+pub mod codegen;
 #[cfg(not(test))]
 pub mod loader;
 #[cfg(not(test))]
