+++ release-64bit-windows-rs.cmd
@@ -1,5 +1,5 @@
 @echo off
-cargo +nightly build --target x86_64-pc-windows-msvc --bin basm-submit --release || goto :error
+cargo +nightly build --target x86_64-pc-windows-msvc --bin basm-submit --features=submit --release || goto :error
 python scripts/static-pie-gen.py basm/src/solution.rs x86_64-pc-windows-msvc target/x86_64-pc-windows-msvc/release/basm-submit.exe static-pie-stub-amd64.bin Rust static-pie-template-amd64.rs || goto :error
 
 :; exit 0
