+++ release-wasm32.cmd
@@ -1,5 +1,5 @@
 @echo off
-cargo +nightly build --target wasm32-unknown-unknown --bin=basm-submit --release || goto :error
+cargo +nightly build --target wasm32-unknown-unknown --bin=basm-submit --features=submit --release || goto :error
 python scripts/wasm-gen.py wasm-template.js JavaScript || goto :error
 
 :; exit 0
